slice navigation into smaller components
