+++ src/Navigation/Header/index.js
@@ -1,0 +1,11 @@
+import { Logo, LogoWrapper, Title } from "./styled"
+import logo from "../../assets/logo.svg";
+
+export const Header = () => {
+    return (
+        <LogoWrapper>
+          <Logo src={logo} alt="" />
+          <Title>movies&nbsp;browser</Title>
+        </LogoWrapper>
+    )
+};
